Added proper exception handling for nested dict updater
- Refactored to have it in a separate place (code + tests)
- Added basic documentation
